Move rate limiting to separate file. Add missing requirement for exports.
