+++ harvester/CKANRepository.py
@@ -1,5 +1,5 @@
 from harvester.HarvestRepository import HarvestRepository
-from functools import wraps
+from harvester.rate_limited import rate_limited
 import ckanapi
 import time
 import json
@@ -223,35 +223,7 @@ class CKANRepository(HarvestRepository):
 
         return record
 
-    def _rate_limited(max_per_second):
-        """ Decorator that make functions not be called faster than a set rate """
-        threading = __import__('threading')
-        lock = threading.Lock()
-        min_interval = 1.0 / float(max_per_second)
-
-        def decorate(func):
-            last_time_called = [0.0]
-
-            @wraps(func)
-            def rate_limited_function(*args, **kwargs):
-                lock.acquire()
-                elapsed = time.clock() - last_time_called[0]
-                left_to_wait = min_interval - elapsed
-
-                if left_to_wait > 0:
-                    time.sleep(left_to_wait)
-
-                lock.release()
-
-                ret = func(*args, **kwargs)
-                last_time_called[0] = time.clock()
-                return ret
-
-            return rate_limited_function
-
-        return decorate
-
-    @_rate_limited(5)
+    @rate_limited(5)
     def _update_record(self, record):
         # self.logger.debug("Updating CKAN record {}".format(record['local_identifier']) )
