+++ harvester/MarkLogicRepository.py
@@ -1,5 +1,4 @@
 from harvester.HarvestRepository import HarvestRepository
-from functools import wraps
 import requests
 import time
 import json
